+++ ricaun.RevitTest.TestAdapter/TestCaseUtils.cs
@@ -54,26 +54,17 @@ namespace ricaun.RevitTest.TestAdapter
 
         private static void SplitTestName(string testName, out string fullyQualifiedName, out string displayName)
         {
-            var splitDots = testName.Split('.');
-
-            var index = LastIndexSplitOfDisplayName(splitDots);
-
+            string[] splitDots = System.Text.RegularExpressions.Regex.Split(testName, pattern);
+            var index = splitDots.Length - 1;
             fullyQualifiedName = string.Join(".", splitDots.Take(index));
             displayName = string.Join(".", splitDots.Skip(index));
 
-            // AdapterLogger.Logger.DebugOnlyLocal($"SplitTestName[{index}]: {testName} -\t {fullyQualifiedName}\t {displayName}");
-        }
+            //AdapterLogger.Logger.DebugOnlyLocal($"SplitTestName[{index}]: {testName} -\t {fullyQualifiedName}\t {displayName}");
 
-        private static int LastIndexSplitOfDisplayName(string[] splitDots)
-        {
-            for (int i = 0; i < splitDots.Length; i++)
-            {
-                var name = splitDots[i];
-                if (name.Contains('(') | name.Contains('"'))
-                    return i;
-            }
-            return splitDots.Length - 1;
+            ///https://github.com/nunit/nunit3-vs-adapter/blob/master/src/NUnitTestAdapter/TestConverter.cs#L78
         }
+        static string pattern = @"(?<!\([^\)]*)\.(?![^\(]*\))";
+
         #endregion
     }
 }